Update
Thanks to Prof. Kazman The label now changes correctly. I missed some simple HTML errors.

Need to fix

the /check function because it is letting the user get to the cart even with an invalid quantity.

And data transferring from the product_display to the cart and then to the receipt